backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow-cpu==2.14.0
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
gunicorn==21.2.0